+++ changelog.md
@@ -13,19 +13,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Typing for the returned values in the methods.
 - file utils.py with supporting logic.
 - simple CI via GitHub Actions.
+- file exception.py with extended errors.
+- file interactive.py with logic in interactive mode.
+- file arguments.py with logic in CLI mode.
 
 ### Changed
 - Naming of methods.
+- decoupling logic from main.py file to two separate files interactive.py and arguments.py.
 
 ### Fixed
 - Some grammatical issues and typos.
 - Rounding input.
+- Possibility provide the temperature below than absolute zero.
 
 
 ## [1.0.1](https://github.com/MariaMaximova/convert_temp/tree/v1.0.1) - 2022 - August - 10
 ### Added
 - *setup.py* file.
-- Added separate folder *converter* within file *converter.py* with all convertion logic.
+- Added separate folder *converter* within file *converter.py* with all conversion logic.
 
 ### Removed
 - Convert logic in the main.py file.